srp-suite-main/agents/nlm-agent/pyproject.toml
[project]
name = "nlm-agent"
version = "0.1.0"
description = "Autonomous NotebookLM MCP agent powered by Gemini API"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "mcp>=1.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
nlm-agent = "nlm_agent.main:main"
